+++ src/main/java/online/kingdomkeys/kingdomkeys/block/GhostBloxBlock.java
@@ -58,8 +58,8 @@ public class GhostBloxBlock extends BaseBlock implements SimpleWaterloggedBlock 
         if (!state.getValue(VISIBLE) && (!state.getValue(BlockStateProperties.WATERLOGGED) && fluidStateIn.getType() == Fluids.WATER)) {
             if (!worldIn.isClientSide()) {
                 worldIn.setBlock(pos, state.setValue(BlockStateProperties.WATERLOGGED, Boolean.valueOf(true)), 3);
-                //TODO figure out this
-                //worldIn.getFluidTicks().scheduleTick(pos, fluidStateIn.getType(), fluidStateIn.getType().getTickDelay(worldIn));
+                // Waterlogged
+                worldIn.scheduleTick(pos, fluidStateIn.getType(), fluidStateIn.getType().getTickDelay(worldIn));
             }
 
             return true;