+++ src/main/java/online/kingdomkeys/kingdomkeys/lib/PortalData.java
@@ -15,11 +15,10 @@ public class PortalData {
     String name;
     BlockPos pos;
     ResourceKey<Level> dimKey;
-    //TODO fix casting
     public PortalData(UUID id, String name, double x, double y, double z, ResourceKey<Level> dimID, UUID ownerID) {
     	this.uuid = id;
     	this.name = name;
-    	this.pos = new BlockPos((int) x, (int) y, (int) z);
+    	this.pos = BlockPos.containing(x, y, z);
         this.dimKey = dimID;
         this.ownerID = ownerID;
     }
@@ -59,12 +58,7 @@ public class PortalData {
         return ownerID;
     }
     public void setOwnerID(UUID id) {
-        this.ownerID = id;
-    } 
-
-    public String getShortCoords() {
-        //return x+", "+y+", "+z;
-        return name;
+	    this.ownerID = id;
     }
     
     public CompoundTag write() {
@@ -83,8 +77,7 @@ public class PortalData {
 		this.setUUID(nbt.getUUID("uuid"));
 		this.setName(nbt.getString("name"));
 
-        //TODO fix casting
-		this.setPos(new BlockPos((int) nbt.getDouble("x"), (int) nbt.getDouble("y"), (int) nbt.getDouble("z")));
+		this.setPos(BlockPos.containing(nbt.getDouble("x"), nbt.getDouble("y"), nbt.getDouble("z")));
 		ResourceLocation rl = KingdomKeys.rl(nbt.getString("dim"));
 		this.setDimID(ResourceKey.create(Registries.DIMENSION,rl));
 		this.setOwnerID(nbt.getUUID("owner"));